+++ tests/spec/return_tests.cpp
@@ -25,7 +25,7 @@ BACKEND_TEST_CASE( "Testing wasm <return_0_wasm>", "[return_0_wasm_tests]" ) {
    CHECK(!bkend.call_with_return("env", "type-f32"));
    CHECK(!bkend.call_with_return("env", "type-f64"));
    CHECK(bkend.call_with_return("env", "type-i32-value")->to_ui32() == UINT32_C(1));
-   CHECK(bkend.call_with_return("env", "type-i64-value")->to_ui64() == UINT32_C(2));
+   CHECK(bkend.call_with_return("env", "type-i64-value")->to_ui64() == UINT64_C(2));
    CHECK(bit_cast<uint32_t>(bkend.call_with_return("env", "type-f32-value")->to_f32()) == UINT32_C(1077936128));
    CHECK(bit_cast<uint64_t>(bkend.call_with_return("env", "type-f64-value")->to_f64()) == UINT64_C(4616189618054758400));
    CHECK(!bkend.call_with_return("env", "nullary"));
@@ -45,10 +45,10 @@ BACKEND_TEST_CASE( "Testing wasm <return_0_wasm>", "[return_0_wasm_tests]" ) {
    CHECK(!bkend.call_with_return("env", "as-br_if-cond"));
    CHECK(bkend.call_with_return("env", "as-br_if-value")->to_ui32() == UINT32_C(8));
    CHECK(bkend.call_with_return("env", "as-br_if-value-cond")->to_ui32() == UINT32_C(9));
-   CHECK(bkend.call_with_return("env", "as-br_table-index")->to_ui64() == UINT32_C(9));
+   CHECK(bkend.call_with_return("env", "as-br_table-index")->to_ui64() == UINT64_C(9));
    CHECK(bkend.call_with_return("env", "as-br_table-value")->to_ui32() == UINT32_C(10));
    CHECK(bkend.call_with_return("env", "as-br_table-value-index")->to_ui32() == UINT32_C(11));
-   CHECK(bkend.call_with_return("env", "as-return-value")->to_ui64() == UINT32_C(7));
+   CHECK(bkend.call_with_return("env", "as-return-value")->to_ui64() == UINT64_C(7));
    CHECK(bkend.call_with_return("env", "as-if-cond")->to_ui32() == UINT32_C(2));
    CHECK(bkend.call_with_return("env", "as-if-then", UINT32_C(1), UINT32_C(6))->to_ui32() == UINT32_C(3));
    CHECK(bkend.call_with_return("env", "as-if-then", UINT32_C(0), UINT32_C(6))->to_ui32() == UINT32_C(6));
@@ -70,14 +70,14 @@ BACKEND_TEST_CASE( "Testing wasm <return_0_wasm>", "[return_0_wasm_tests]" ) {
    CHECK(bkend.call_with_return("env", "as-local.tee-value")->to_ui32() == UINT32_C(1));
    CHECK(bkend.call_with_return("env", "as-global.set-value")->to_ui32() == UINT32_C(1));
    CHECK(bit_cast<uint32_t>(bkend.call_with_return("env", "as-load-address")->to_f32()) == UINT32_C(1071225242));
-   CHECK(bkend.call_with_return("env", "as-loadN-address")->to_ui64() == UINT32_C(30));
+   CHECK(bkend.call_with_return("env", "as-loadN-address")->to_ui64() == UINT64_C(30));
    CHECK(bkend.call_with_return("env", "as-store-address")->to_ui32() == UINT32_C(30));
    CHECK(bkend.call_with_return("env", "as-store-value")->to_ui32() == UINT32_C(31));
    CHECK(bkend.call_with_return("env", "as-storeN-address")->to_ui32() == UINT32_C(32));
    CHECK(bkend.call_with_return("env", "as-storeN-value")->to_ui32() == UINT32_C(33));
    CHECK(bit_cast<uint32_t>(bkend.call_with_return("env", "as-unary-operand")->to_f32()) == UINT32_C(1079613850));
    CHECK(bkend.call_with_return("env", "as-binary-left")->to_ui32() == UINT32_C(3));
-   CHECK(bkend.call_with_return("env", "as-binary-right")->to_ui64() == UINT32_C(45));
+   CHECK(bkend.call_with_return("env", "as-binary-right")->to_ui64() == UINT64_C(45));
    CHECK(bkend.call_with_return("env", "as-test-operand")->to_ui32() == UINT32_C(44));
    CHECK(bkend.call_with_return("env", "as-compare-left")->to_ui32() == UINT32_C(43));
    CHECK(bkend.call_with_return("env", "as-compare-right")->to_ui32() == UINT32_C(42));
